LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-form-element/form-elements-matches-expected.txt

 

PASS input type=image should not be present in the form.elements collection 
FAIL form.elements should include elements whose name starts with a number assert_equals: [2] expected (undefined) undefined but got (object) Element node <input name="2"></input>
